+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Update field `updated_dt` on queryset update (OSIDB-3573)
 - Introduce purl to Affect (OSIDB-3409)
+- Implement field `embargoed` for advanced search (OSIDB-3549)
 
 ## [4.5.6] - 2024-11-08
 ### Fixed